+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # webthing Changelog
 
+## [Unreleased]
+### Added
+- Ability to set a base URL path on server.
+
 ## [0.11.0] - 2019-01-16
 ### Changed
 - `WebThingServer::new()` can now take a configuration function which can add additional API routes.