+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - Ability to set a base URL path on server.
+### Changed
+- Things now use `title` rather than `name`.
 
 ## [0.11.0] - 2019-01-16
 ### Changed